requirements.txt
numpy==2.2.1
opencv_python_headless==4.10.0.84
Pillow==11.1.0
pycryptodome==3.20.0
PyYAML==6.0.2
Requests==2.32.3
schedule==1.2.2
